- got rid of small integer primitives byte and short (again :D). they saved a bit of a RAM, but the whole code was obstructed with type conversions

+++ src/main/java/emulator/AddressRegister.java
@@ -1,6 +1,6 @@
 /*
  * Developed by Patrick Metz <patrickmetz@web.de>.
- * Last modified 25.02.19 18:37.
+ * Last modified 01.03.19 15:22.
  * Copyright (c) 2019. All rights reserved.
  */
 
@@ -8,13 +8,13 @@ package emulator;
 
 final class AddressRegister {
 
-    private short memory;
+    private int memory;
 
-    short read() {
+    int read() {
         return memory;
     }
 
-    void write(short value) {
+    void write(int value) {
         memory = value;
     }
