+++ src/main/java/emulator/DataRegisters.java
@@ -1,6 +1,6 @@
 /*
  * Developed by Patrick Metz <patrickmetz@web.de>.
- * Last modified 25.02.19 18:37.
+ * Last modified 01.03.19 15:23.
  * Copyright (c) 2019. All rights reserved.
  */
 
@@ -10,17 +10,17 @@ import java.util.HashMap;
 
 final class DataRegisters {
 
-    private final HashMap<Byte, Byte> memory;
+    private final HashMap<Integer, Integer> memory;
 
     DataRegisters() {
         memory = new HashMap<>(16, 2);
     }
 
-    byte read(byte address) {
+    Integer read(int address) {
         return memory.get(address);
     }
 
-    void write(byte address, byte value) {
+    void write(int address, int value) {
         memory.put(address, value);
     }
