+++ src/main/java/emulator/Graphics.java
@@ -1,6 +1,6 @@
 /*
  * Developed by Patrick Metz <patrickmetz@web.de>.
- * Last modified 26.02.19 12:46.
+ * Last modified 01.03.19 15:40.
  * Copyright (c) 2019. All rights reserved.
  */
 
@@ -8,11 +8,11 @@ package emulator;
 
 final class Graphics {
 
-    private static final byte SCREEN_HEIGHT = 32;
-    private static final byte SCREEN_WIDTH = 64;
-    private static final byte SPRITE_WIDTH = 8;
+    private static final int SCREEN_HEIGHT = 32;
+    private static final int SCREEN_WIDTH = 64;
+    private static final int SPRITE_WIDTH = 8;
 
-    private byte[][] screen;
+    private int[][] screen;
 
     Graphics() {
         initializeScreen();
@@ -22,15 +22,15 @@ final class Graphics {
         initializeScreen();
     }
 
-    boolean drawSprite(byte offsetX, byte offsetY, byte[] rows) {
+    boolean drawSprite(int offsetX, int offsetY, int[] rows) {
         boolean collision = false;
 
-        for (byte x = 0; x < SPRITE_WIDTH; x++) {
-            for (byte y = 0; y < rows.length; y++) {
-                byte newPixelState = (byte) ((rows[y] >> x) & 1);
+        for (int x = 0; x < SPRITE_WIDTH; x++) {
+            for (int y = 0; y < rows.length; y++) {
+                int newPixelState = (rows[y] >> x) & 1;
 
-                byte positionX = (byte) ((offsetX + x) % SCREEN_WIDTH);
-                byte positionY = (byte) ((offsetY + y) % SCREEN_HEIGHT);
+                int positionX = (offsetX + x) % SCREEN_WIDTH;
+                int positionY = (offsetY + y) % SCREEN_HEIGHT;
 
                 if ((newPixelState == 1) && (screen[positionX][positionY] == 1)) {
                     newPixelState = 0;
@@ -44,12 +44,12 @@ final class Graphics {
         return collision;
     }
 
-    byte[][] getScreen() {
+    int[][] getScreen() {
         return screen;
     }
 
     private void initializeScreen() {
-        screen = new byte[SCREEN_WIDTH][SCREEN_HEIGHT];
+        screen = new int[SCREEN_WIDTH][SCREEN_HEIGHT];
     }
 
 }
