+++ src/main/java/emulator/Keyboard.java
@@ -1,6 +1,6 @@
 /*
  * Developed by Patrick Metz <patrickmetz@web.de>.
- * Last modified 28.02.19 23:29.
+ * Last modified 01.03.19 17:29.
  * Copyright (c) 2019. All rights reserved.
  */
 
@@ -10,30 +10,30 @@ import java.util.HashMap;
 
 final class Keyboard extends Thread {
 
-    private final HashMap<Byte, Boolean> keys;
+    private final HashMap<Integer, Boolean> keys;
 
     Keyboard() {
         // 16 hexadecimal key codes, from 0 to F
         keys = new HashMap<>(16, 2);
 
         for (int i = 0; i < 0xF; i++) {
-            keys.put((byte) i, false);
+            keys.put(i, false);
         }
     }
 
-    public void pressKey(byte keyCode) {
+    public void pressKey(int keyCode) {
         keys.put(keyCode, true);
     }
 
-    public void releaseKey(byte keyCode) {
+    public void releaseKey(int keyCode) {
         keys.put(keyCode, false);
     }
 
-    boolean isKeyPressed(byte keyCode) {
+    boolean isKeyPressed(int keyCode) {
         return keys.get(keyCode);
     }
 
-    byte waitForKey() {
+    Integer waitForKey() {
         return 0xF; // just fake it for now ;-P
     }
